requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
github-copilot-sdk>=0.1.0
python-dotenv>=1.0.0
soundfile>=0.12.0
numpy>=1.24.0
